refactor classify VS Code adapter layout
