+++ src/bootstrap/extension/extensionSubscriptions.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
-import { registerDiagnostics } from "../diagnostics/registerDiagnostics";
-import { registerHoverProvider } from "../languages/registerHoverProvider";
+import { registerDiagnostics } from "../../presentation/vscode/diagnostics/registerDiagnostics";
+import { registerHoverProvider } from "../../presentation/vscode/languages/registerHoverProvider";
 import type { ExtensionDependencies } from "./extensionDependencies";
 import { createWebApiImportSubscriptions } from "./webapiImportWiring";
 import { createViewerSubscriptions } from "./viewerWiring";
