+++ src/infrastructure/telemetry/NoopTelemetryAdapter.ts
