+++ src/infrastructure/telemetry/VscodeTelemetryAdapter.ts
