+++ src/presentation/vscode/common/texteditorFinder.ts
