+++ src/test/suite/viewerWiring.test.ts
@@ -1,7 +1,7 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
 import type { TelemetryPort } from "../../application/telemetry/TelemetryPort";
-import { createViewerSubscriptions } from "../../extension/bootstrap/viewerWiring";
+import { createViewerSubscriptions } from "../../bootstrap/extension/viewerWiring";
 
 suite("Viewer wiring", () => {
   test("creates viewer subscriptions for both table and flow viewers", () => {
