+++ tsconfig.test.json
@@ -21,7 +21,7 @@
     "src/domain/models/ajs/**/*.ts",
     "src/domain/services/parser/**/*.ts",
     "src/domain/values/**/*.ts",
-    "src/extension/telemetry/**/*.ts",
+    "src/infrastructure/telemetry/**/*.ts",
     "src/generate/parser/**/*.ts"
   ]
 }
